ref: Background is not a layout
