+++ src/elements/background.ts
@@ -1,5 +1,24 @@
-import { BoundingBox, Context, Element } from "../base";
+import { BoundingBox, Context, Element, Drawable } from "../base";
 import { ColorValue, getRGB } from "../utils";
+
+/**
+ * `Background` is used to draw a background color over any
+ * arbitrary bounding box.
+ */
+export class Background implements Drawable {
+  /**
+   * Create a background that can be drawn anywhere
+   * @param color background color
+   */
+  constructor(private color: ColorValue) {}
+
+  draw(context: Context, box: BoundingBox): void {
+    context.raw
+      .rect(box.x, box.y, box.width, box.height)
+      .fill(getRGB(this.color));
+  }
+}
+
 /**
  * `ElementBackground` is basically a background drawn into the bounding
  * box of any element.
